Gets collection view kinda working
