--- Atletas/ExploreContent.swift
@@ -1,15 +1,0 @@
-//
-//  ExploreContent.swift
-//  Atletas
-//
-//  Created by André Schueda on 02/09/21.
-//
-
-import UIKit
-
-class ExploreContent: UIView {
-
-    lazy var exploreCollection = {
-        
-    }()
-}
