+++ Atletas/ExploreView/ExploreViewController.swift
@@ -1,0 +1,84 @@
+//
+//  ExploreViewController.swift
+//  Atletas
+//
+//  Created by Djenifer Renata Pereira on 26/08/21.
+//
+
+import UIKit
+import SnapKit
+
+class ExploreViewController: UIViewController {
+    
+    lazy var searchBar: UISearchBar = {
+        let searchBar = UISearchBar()
+        searchBar.translatesAutoresizingMaskIntoConstraints = false
+        searchBar.searchBarStyle = .minimal
+        searchBar.placeholder = "Procurar"
+        return searchBar
+    }()
+    
+    lazy var contentView: ExploreContentView = {
+        let contentView = ExploreContentView()
+        contentView.translatesAutoresizingMaskIntoConstraints = false
+        return contentView
+    }()
+    
+    override func loadView() {
+        super.loadView()
+        
+        setupSearchBar()
+        setupContent()
+    }
+    
+    func setupSearchBar() {
+        view.addSubview(searchBar)
+        searchBar.snp.makeConstraints { make in
+            make.top.equalTo(self.view.safeAreaLayoutGuide.snp.top)
+            make.leading.equalToSuperview().offset(8)
+            make.trailing.equalToSuperview().offset(-8)
+        }
+    }
+    
+    func setupContent() {
+        view.addSubview(contentView)
+        contentView.snp.makeConstraints { make in
+            make.top.equalTo(searchBar.snp.bottom)
+            make.leading.equalToSuperview()
+            make.trailing.equalToSuperview()
+            make.bottom.equalTo(self.view.safeAreaLayoutGuide.snp.bottom)
+        }
+    }
+    
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        view.backgroundColor = .white
+        navigationController?.navigationBar.prefersLargeTitles = true
+        navigationItem.title = "Ajude um atleta"
+        
+    }
+    
+    
+}
+
+#if DEBUG
+import SwiftUI
+struct ViewController_Preview: PreviewProvider {
+    static var previews: some View {
+        ContentView().previewDevice("iPhone 12")
+    }
+
+    struct ContentView: UIViewControllerRepresentable {
+
+        func makeUIViewController(context: Context) -> UIViewController {
+            UINavigationController(rootViewController: ExploreViewController())
+        }
+
+        func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
+        }
+    }
+}
+#endif
